+++ HISTORY.rst
@@ -9,6 +9,7 @@ History
 * Add an API client to wrap the Data for SEO RestClient
 * Add ability to get credentials from environment variables
 ** DFS_API_USERNAME and DFS_API_PASSWORD
+* Add feature to save requests as files
 * Add service to get locations and languages
 * Add service to get domain competitors
-* Add feature to save requests as files
+* Add service to get ranked keywords for a domain